apps/geocore-api/requirements.txt
amqp==2.5.2
APScheduler==3.6.3
asgiref==3.2.3
billiard==3.6.3.0
celery==4.4.2
certifi==2019.11.28
chardet==3.0.4
coverage==4.5.1
Django==3.0.4
django-appconf==1.0.4
django-cors-headers==3.2.1
django-debug-toolbar==2.2
django-filter==1.1.0
djangorestframework==3.11.0
djangorestframework-filters==0.11.1
djangorestframework-gis==0.15
docutils==0.16
idna==2.9
importlib-metadata==1.6.0
jsonfield==3.1.0
kombu==4.6.8
model-mommy==1.5.1
psycopg2-binary==2.8.4
pylru==1.2.0
python-dotenv==0.12.0
pytz==2019.3
requests==2.23.0
six==1.14.0
sqlparse==0.3.1
tzlocal==2.0.0
urllib3==1.25.8
vine==1.3.0
zipp==3.1.0
